+++ example/main.dart
@@ -1,7 +1,7 @@
 import 'package:tiengviet/tiengviet.dart';
 
 void main() {
-  String str = tiengviet("Hello Việt Nam");
+  final str = TiengViet.parse('Hello Việt Nam');
   //ignore: avoid_print
   print(str); // Hello Viet Nam
 }